requirements.txt
celery==5.2.6
Django==3.2.9
django_celery_beat==2.2.1
djangorestframework==3.13.1
pydantic==1.9.0
python_dateutil==2.8.2
super_logger==0.1.4
